refactor(discovery): extract ADR Discovery into its own top-level module
Discovery was never part of the Sensor. It shared a package only because that
is where it was first written: nothing under discovery/ imported from
adr_sensor, and nothing in adr_sensor reached into discovery except one
subcommand dispatch in the CLI. The two planes have different dependencies and
different deployment stories, and the repository layout now says so.

Sensor/adr_sensor/discovery/ becomes Discovery/adr_discovery/, and
Sensor/tests/discovery/ becomes Discovery/tests/. The package ships its own
pyproject with an adr-discovery console script in place of `adr-sensor
discover`. The tomli dependency and the catalog.json package-data entry move
with it, because both existed only for discovery and the Sensor wheel should
not carry either.

With discovery gone the lazy-import scaffolding in adr_sensor is dead. Its
whole purpose was to let the collector run on an endpoint that has only the
standard library, without dragging in tabulate by way of .observer. The package
__init__ and the CLI both go back to importing the observability plane
directly.

CI gains a Discovery job on the same 3.9 through 3.13 matrix as the Sensor,
behind a stable "Discovery tests" check name.

Nothing about the collector's behaviour changed. 490 fidelity checks pass
across 241 cases from the new location, the Sensor's 123 tests are unaffected,
both packages lint clean, a live scan still reports 31 assets with no errors,
and the built wheel contains catalog.json.
