+++ Discovery/adr_discovery/probes/cli_agent.py
