+++ Discovery/tests/cases_skills.py
