+++ Discovery/tests/test_fidelity.py
