+++ README.md
@@ -8,7 +8,7 @@ ADR is **deployed in production at Uber**, and the accompanying paper was accept
 
 ADR secures enterprise AI agents through five complementary capabilities: discovering unsanctioned AI tools, observing agent activity, evaluating defenses, detecting threats, and preventing unsafe actions.
 
-1. **ADR Discovery: Find the AI tools present on employee endpoints.** Inventories installed AI applications, CLI agents, local model runtimes, and MCP servers. This component is not included in the current open-source release. **Stay tuned.**
+1. **ADR Discovery: Find the AI tools present on employee endpoints.** Inventories installed AI applications, CLI agents, IDE extensions, local model runtimes, and MCP servers, and flags unknown surfaces for review.
 2. **ADR Observability: Understand what AI agents are doing and why.** In production, ADR captures agent intent, tool use, and execution traces across 7+ AI coding tools on macOS, Linux, and Windows, as well as internal automation and customer-facing support agents.
 3. **ADR Benchmark: Test agent security under realistic enterprise conditions.** ADR-Bench includes 300+ tasks, 133 MCP servers, and coverage of all 17 agent attack techniques.
 4. **ADR Detection: Detect risky agent behavior efficiently.** Its two-tier architecture combines high-recall triage with deeper agentic reasoning for suspicious sessions.
@@ -16,10 +16,11 @@ ADR secures enterprise AI agents through five complementary capabilities: discov
 
 ## Repository layout
 
-This repository contains the open-source **ADR Sensor**, **ADR-Bench**, and **ADR Detector** described in the paper. The offline **ADR Explorer** engine, which hardens ADR Detection through pre-deployment red teaming, is not included here.
+This repository contains the open-source **ADR Discovery**, **ADR Sensor**, **ADR-Bench**, and **ADR Detector** described in the paper. The offline **ADR Explorer** engine, which hardens ADR Detection through pre-deployment red teaming, is not included here.
 
 | Path                                               | ADR component              | Description                                                                          |
 | -------------------------------------------------- | -------------------------- | ------------------------------------------------------------------------------------ |
+| [Discovery/](Discovery/)                           | ADR Discovery              | Inventory the AI apps, CLI agents, IDE extensions, model runtimes, and MCP servers on an endpoint, and flag unknown surfaces for review |
 | [Sensor/](Sensor/)                                 | ADR Observability          | Collect and normalize agent telemetry from Claude Code, Cursor, Codex, opencode, Claude Desktop, and others |
 | [Detection/](Detection/)                           | ADR Benchmark + Detection  | Dual-agent detector, 133 MCP servers, 303 benchmark tasks, baselines, figure scripts |
 | [docs/REPRODUCIBILITY.md](docs/REPRODUCIBILITY.md) | Evaluation                 | Step-by-step workflow to reproduce benchmark detection and paper figures             |
@@ -39,6 +40,7 @@ See **[docs/REPRODUCIBILITY.md](docs/REPRODUCIBILITY.md)** for the full evaluati
 
 Component documentation:
 
+- [Discovery/README.md](Discovery/README.md): endpoint inventory, probes, and the fingerprint catalog
 - [Sensor/README.md](Sensor/README.md): telemetry collection and unified schema
 - [Detection/README.md](Detection/README.md): ADR-Bench, detector baselines, MCP infrastructure
